feat: organized recipe cards
Renamed the old RecipeCard -> SavedRecipeCard
Broke out some parts of the RecipeCard, like Wrapper, Image and StarRating
