+++ frontend/src/components/Dashboard.tsx
@@ -1,6 +1,6 @@
 import { Box, Flex, Heading, List, ListItem, Checkbox } from '@chakra-ui/react';
 
-import RecipeCard from './RecipeCard';
+import SavedRecipeCard from './SavedRecipeCard';
 
 import { RECIPES } from '@/data';
 
@@ -28,7 +28,7 @@ const Dashboard = () => {
             Your Recipes
           </Heading>
           {RECIPES.map((recipe, index) => (
-            <RecipeCard key={index} recipe={recipe} />
+            <SavedRecipeCard key={index} recipe={recipe} />
           ))}
         </Box>
